+++ CHANGELOG.md
@@ -10,7 +10,7 @@
 * DLT Debug mode when source is streaming and node is not
 * DataFrame type propagation when models are used as inputs to other models
 ### Updated
-* n/a
+* Changed default value for mergeSchema to `False` when mode is `COMPLETE` or `OVERWRITE`
 ### Breaking changes
 * n/a