+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ### Fixed
 * DLT Debug mode when source is streaming and node is not
 * DataFrame type propagation when models are used as inputs to other models
+* `show_version_info()` method to display correct packages version
 ### Updated
 * Changed default value for mergeSchema to `False` when mode is `COMPLETE` or `OVERWRITE`
 ### Breaking changes